fixat addChore
